core: use number of endpoints as primary port type hint
If a port's USB interface has three USB endpoints, it's almost
certain they are going to be a bulk-in, bulk-out, and an interrupt
endpoint.  The presence of an interrupt endpoint usually signals
that the port can support PPP and thus should likely be our data
port.  Use number of endpoints as a hint in addition to udev rules
to hopefully get probing correct on a wider range of modems that
we haven't seen before.