broadband-modem: prefer QCDM signal quality checks for CDMA-only modems
Now that we can get EVDO signal strength too, just use QCDM if we have
it.  For CDMA devices +CSQ is ambiguous because it almost always reports
only CDMA1x signal strength, not EVDO strength.  So attempting to use
AT commands for signal strength always returns the wrong result when
the device is using EVDO.