serial: split 'MMSerialPort' and subclass a new 'MMCommandSerialPort'
We take out from the 'MMSerialPort' all the logic related to sending and
receiving commands, and we add it to a new 'MMCommandSerialPort'. All AT, QCDM
and GPS port objects will now inherit from this new 'MMCommandSerialPort'.

We do this change in order to be able to subclass a 'MMSerialPort' without any
command-related logic.