sim: if no MNC length in EFad, guess it from the MCC value
Quoting 3GPP TS 23.003:
   IMSI is composed of three parts:
   i) Mobile Country Code (MCC) consisting of three digits. The MCC identifies
      uniquely the country of domicile of the mobile subscriber;
   ii) Mobile Network Code (MNC) consisting of two or three digits for GSM
       applications. The MNC identifies the home GSM PLMN of the mobile
       subscriber. The length of the MNC (two or three digits) depends on the
       value of the MCC. A mixture of two and three digit MNC codes within a
       single MCC area is not recommended and is outside the scope of this
       specification.
   iii) Mobile Subscriber Identification Number (MSIN) identifying the mobile
        subscriber within a GSM PLMN

We can therefore assume that there are specific MCCs where the MNC will be 3
digits long, instead of 2.

Using this logic, and information retrieved from the m-b-p-i database, try to
guess the MNC length if the EFad file is not readable or if it doesn't have the
MNC length field; and use it to construct the MCC/MNC string.