iface-modem-location: interface context is ref-counted
We're going to add a mutex within the context, so we need to make sure that we
don't try to clear a locked mutex. Whenever the mutex is locked, a valid
reference to the context will be held, so freeing the context will only happen
when the mutex is unlocked.