introspection,location: new 'LocationUpdated' signal
This signal is emitted whenever the location information is updated in
ModemManager, and instructs clients that they can re-fetch the location
information, e.g. via GetLocation().

When clients are able to use and track the 'Location' property, they don't
really need this signal, as they can follow the changes notified by the
PropertiesChanged signal in the org.freedesktop.DBus.Properties interface.