broadband-modem: allow plugins to specify USSD type (encoded/unencoded) to use
A new "broadband-modem-ussd-type" property is included in the MMBroadbandModem
object, which allows subclasses of this type to specify whether they want to
restrict USSD to encoded or unencoded formats. If UNKNOWN given (the default),
the logic will fallback to the previous logic, where first encoded is tried,
and then unencoded is tried.