base-modem: remove port from the modem and explicitly dispose it
The port wasn't being removed from the modem object when we got it notified via
udev, just from the MMDevice object. The logic was anyway working because once
the last port was removed from the MMDevice object, the full modem object was
marked as invalid and removed itself.

Instead, we now remove ports one by one when we get notified about them, and we
also will explicitly run dispose() on them, to e.g. have AT ports marked as
forced-close or QMI ports with all clients released. This will make any next
operation get a direct error and hence a quicker full removal. In particular,
unplugging a modem which is in connected state should not end up trying to open
the already gone AT port, as we had in GB#701809:

    unexport_modem(): Unexported modem '/sys/devices/virtual/tty/rfcomm0' from path '/org/freedesktop/ModemManager1/Modem/3'
    mm_port_serial_open(): (rfcomm0) opening serial port...
    mm_port_serial_open(): (rfcomm0) could not open serial device (2)
    disconnect_ready(): Couldn't disconnect bearer '/org/freedesktop/ModemManager1/Bearer/1'
     _close_internal: assertion 'self->priv->open_count > 0' failed
    port_serial_close_force(): (rfcomm0) forced to close port
    finalize(): Modem (Generic) '/sys/devices/virtual/tty/rfcomm0' completely disposed

https://bugzilla.gnome.org/show_bug.cgi?id=701809